+++ packages/alfa-promise/src/builtin.ts
@@ -1,4 +1,4 @@
-// This file defines exports from the global `Promise` constructor for internal
+// This file defines exports from the builtin `Promise` constructor for internal
 // use only.
 
 /**